backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django==5.1",
    "django-bootstrap5>=25.1",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "django-cors-headers",
    "djangorestframework-simplejwt[token_blacklist]",
]
